"Added ActiveProfiles in test classes and removed secret keys from properties file
The ActiveProfiles were added to specify which profiles are active when running the tests in the 'DetectionControllerTest', 'ChartControllerIntegrationTest', 'CameraControllerTest', 'ImageControllerTest', 'LiveViewControllerTest', 'AlarmControllerTest', and 'FirebaseControllerTest' Kotlin test classes. This will allow testing under various application configurations. This change is necessary for creating an environment specific to testing.

Additionally, the 'keycloak.credentials.secret' and 'spring.security.oauth2.client.registration.banyuwangi-test.client-secret' keys were removed from the application.properties file to improve security by avoiding the exposure of secret keys. Also, the 'streaming.streamingToken' had been deleted because it is not required anymore in this configuration."